remove duplicate schemas, reorganize files
